backend/requirements.txt
django
django_filter==23.3
djoser==2.2.0
python-dotenv==1.0.0
djangorestframework==3.14.0
drf-extra-fields==3.7.0
Pillow==10.0.1
gunicorn==21.2.0
psycopg2-binary==2.9.9
coreschema==0.0.4
django-cors-headers==3.13.0